requirements.txt
beautifulsoup4==4.13.4
requests==2.32.4
python-dateutil==2.9.0.post0
langchain==0.3.25
langchain-community==0.3.25
langchain-google-genai==2.1.5
langchain-experimental==0.3.4
langchain-neo4j==0.4.0
langchain-pinecone==0.2.8
python-dotenv==1.0.1
openai==1.90.0
streamlit==1.32.2
fastapi==0.115.13
uvicorn==0.34.3
httpx==0.28.1
pydantic==2.11.7
pinecone-client==3.0.0
pyreadline3==3.5.4
selenium==4.33.0
playwright==1.53.0
webdriver-manager==4.0.2
